+++ packages/node-api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @verdaccio/node-api
 
+## 7.0.0-next-7.15
+
+### Minor Changes
+
+- bd8703e: feat: add migrateToSecureLegacySignature and remove enhancedLegacySignature property
+
+### Patch Changes
+
+- Updated dependencies [bd8703e]
+  - @verdaccio/core@7.0.0-next-7.15
+  - @verdaccio/config@7.0.0-next-7.15
+  - @verdaccio/server-fastify@7.0.0-next-7.15
+  - @verdaccio/server@7.0.0-next-7.15
+  - @verdaccio/logger@7.0.0-next-7.15
+
 ## 7.0.0-next-7.14
 
 ### Patch Changes