+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @verdaccio/utils
 
+## 7.0.0-next-7.15
+
+### Patch Changes
+
+- Updated dependencies [bd8703e]
+  - @verdaccio/core@7.0.0-next-7.15
+
 ## 7.0.0-next-7.14
 
 ### Patch Changes